+++ landingzones/views.py
@@ -157,7 +157,7 @@ class ZoneCreateMixin(ZoneConfigPluginMixin):
                 ),
                 'description': zone.description,
                 'zone_config': zone.configuration,
-                'dirs': [],
+                'colls': [],
             },
         )